// apps/lab/src/routes/graphql/-sub/06-pagination/__generated__/solutionDemoPostList_query.graphql.ts
export type solutionDemoPostList_query$data = {
  readonly posts: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly author: {
          readonly name: string;
        };
        readonly id: string;
        readonly title: string;
      };
    }>;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "solutionDemoPostList_query";
};
export type solutionDemoPostList_query$key = {
  readonly " $data"?: solutionDemoPostList_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"solutionDemoPostList_query">;
};